backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
charset-normalizer==3.4.2
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
gunicorn==23.0.0
h5py==3.14.0
idna==3.10
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pvlib==0.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.4
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
